mkudffs: Fix printf numeric formats

+++ mkudffs/main.c
@@ -219,9 +219,9 @@ int main(int argc, char *argv[])
 	buf[len] = 0;
 	printf("uuid=%.16s\n", buf);
 
-	printf("blocksize=%u\n", disc.blocksize);
-	printf("blocks=%u\n", disc.head->blocks);
-	printf("udfrev=%x\n", disc.udf_rev);
+	printf("blocksize=%u\n", (unsigned int)disc.blocksize);
+	printf("blocks=%lu\n", (unsigned long int)disc.head->blocks);
+	printf("udfrev=%x\n", (unsigned int)disc.udf_rev);
 
 	if (((disc.flags & FLAG_BRIDGE) && disc.head->blocks < 513) || disc.head->blocks < 281)
 	{
